fix(cli): fix version and improve perf
