+++ .changeset/late-shoes-drum.md
@@ -1,0 +1,6 @@
+---
+'@rsbuild/shared': patch
+'@rsbuild/core': patch
+---
+
+fix(cli): fix version and improve perf
